requirements.txt
tekore==5.4.0
quart==0.19.6
python-dotenv==1.0.1
tortoise-orm==0.21.3
asyncpg==0.29.0
humanize==4.9.0